Move global instance of RuntimePluginLoader out of a virtual module
This allows easily requiring it without the need to generate and somehow make the tooling find the declaration files for virtual module.

While I considered moving the runtime out into its own package, in the process I realized that it'd require all packages that want to reference the runtime to declare the runtime package as a dependency, which would add unnecessary complexity to the dependencies. For that reason it's staying as part of the build tools which every package will almost certainly already depend on. Creating a webpack alias for the runtime instance in the future is an option that might be worth looking into, though it would reduce transparency a little.