Use a hook for determining BuildTarget's externals
Fixes a regression introduced in 9788edc causing the frontend Vue bundles used for SSR to interact with the backend Koa bundle, resulting in the backend being started multiple times and causing all sorts of weirdness and headaches.

Externals uses a hook instead of webpack-chain like everything else because webpack-chain still doesn't support externals and I'm not sure how to implement it without breaking backwards compatibility horribly. See neutrinojs/webpack-chain#222.

This is the first commit in what will hopefully soon become a series of commits for all the stuff I've done in the past week or so.