framework-koa: Split into an entry point and library with runtime plugin support
This also adds the ability to use the Koa framework without needing to create a custom entry point. If no `index.ts` file exists in the backend directory, the included entry point will be used automatically to initialize the app and its routes. Otherwise, the `index.ts` file will be used as the bundle entry point and it will need to start the app by calling the `RuntimePluginLoader.start()` function.