Rework type extensions
Runtime and build extensions need to be separate to avoid importing the build stuff into the runtime or other weirdness like that.

The glue-koa-vue package now uses TypeScript's `<reference types="">` directive to import the types without actually importing the packages. This is due to TypeScript not importing the global module definition augmentations by default unless the package is in the `@types/*` namespace, is included in the `compilerOptions.types` (which overrides the default importing of all `@types/*` packages for module augmentations), or the package is imported somewhere in the project using an `import` statement.

Since Aedris imports build plugins dynamically at runtime, there's no way for the compiler to tell whether that package's global augmentations were indeed applied, so we have to give it a hint. While technically, according to the documentation, the proper way to do it would be using `compilerOptions.types` (as references are supposed to be for "when you’re authoring a `d.ts` file by hand"), I don't like the prospect of things getting problematic in the future due to that config change, not to mention it would require crafting the `tsconfig.json` for each project by hand instead of just linking them all to a single global one (which I'm hoping to do in the future).

Something to consider is the possibility of making our own `@types/aedris` package that would simply import all the types and by extension would cause all of them to be defined in the global namespace, but this feels a bit overkill, might cause errors if some of the modules aren't installed (as the compiler will most likely complain if one of them isn't found), and overall just feels dirty. Oh well, wonders of TypeScript.

References:
https://www.typescriptlang.org/docs/handbook/triple-slash-directives.html#-reference-types-
https://www.typescriptlang.org/docs/handbook/tsconfig-json.html#types-typeroots-and-types