build-tools: Fix local plugins built twice
I'm not sure if this is the right fix, but it feels right. Tasks are gonna be created through the `ToolsManager`, and the `ToolsManager` always rebuilds the local plugins so it can get tasks from them, meaning that the local plugins will have always already been rebuilt when the `BuildTask` is ran and we should be able to always skip this step.